// libs/features/contributor/src/lib/shared/models/coverage-period-wrapper.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
import { CoveragePeriod } from './coverage-period';

/**
 * This class is wrapper class for coerage period
 */
export class CoveragePeriodWrapper {
  periods: CoveragePeriod[] = [];
  currentPeriod: CoveragePeriod = new CoveragePeriod();
}